backend/requirements.txt
aioredis==2.0.1
aiosmtplib==5.0.0
aiomysql==0.2.0
alembic==1.18.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
fastapi-users==15.0.3
fastapi-users-db-sqlalchemy==7.0.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
makefun==1.16.0
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
passlib[bcrypt]==1.7.4
pwdlib==0.3.0
pycparser==2.23
pydantic==2.12.5
python-jose[cryptography]==3.3.0
pydantic-settings==2.1.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
redis==7.1.0
sentry-sdk==2.49.0
SQLAlchemy==2.0.45
starlette==0.50.0
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
